// src/contexts/app-context/action-creators/ActionTypes.enum.ts
export enum ActionTypes {
    ON_REQUEST_START = "ON_REQUEST_START",
    ON_ASSETS_RESPONSE_SUCCESS = "ON_ASSETS_RESPONSE_SUCCESS",
    ON_SEARCH = "ON_SEARCH",
    ON_SELECTED_ASSETS = "ON_SELECTED_ASSETS",
    ON_SORT = "ON_SORT",
    ON_PAGE_CHANGE = "ON_PAGE_CHANGE",
    ON_CHANGE_FETCH_LIMIT = "ON_CHANGE_FETCH_LIMIT",
    ON_RESET_SELECTION = "ON_RESET_SELECTION",
    ON_ASSET_CHECK = "ON_ASSET_CHECK",
    ON_CLOSE_ASSET_PANEL = "ON_CLOSE_ASSET_PANEL"
  }
  